pyproject.toml
[project]
name = "win_stat_tg_bot"
version = "0.1.0"
description = "Telegram-бот для вывода статистики Windows-машины."
requires-python = ">=3.8"
dependencies = [
    "aiogram>=3.4.1,<4.0.0",
    "psutil",
    "python-dotenv"
]

[tool.uv]
# Здесь можно добавить дополнительные настройки uv при необходимости 